@@ -1,6 +1,8 @@
 <?php
 
-class JSTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class JSTest extends TestCase
 {
     protected function assertJSRender($content)
     {
@@ -10,7 +12,7 @@ class JSTest extends PHPUnit_Framework_TestCase
     public static $BEGIN = "<script type=\"text/javascript\">var APP = APP || {'settings': {}, 'behaviors':{}, 'locale':{}, 'utilities':{}};";
     public static $END = '</script>';
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
